Added staticcly declared TRefs (compile time - evaluates to a u32)
TRef Hello( TRef_Static(h,e,l,l,o) );

for shorter refs use TRef_StaticX;

TRef Hell = TRef_Static4(h,e,l,l);
TMessage( TRef_Static1(x) );

etc

Removed the virtual from the TArray GetSize. This was originally virtual for other array types but they all use the m_Size of the base class now. Compiler probably inlined this anyway. 

Added a TBinaryTree copy constructor. This might have saved you a bit of a headache duane :) It looks like the child TPtrArray was blindly copying the TPtr data which is a little odd.

Changed the core update timing recording so that it records the time of all the subscribers of the update to help break down the profiling a little. I've had to expose some subscriber stuff but we can revoke this later (or change it to an explicit Debug_ function)
TSubscriber now has a virtual GetSubscriberRef() function. defaults to nothing but can be overloaded, eg. managers return their manager ref.

Removed the update/update all node traversal of the render graph. Now messages sent directly to nodes are queued up on the graph and dispatched at the graph update rather than in every node's own update. Should be a good saving as the traversal was quite slow and ended up doing barely anything. May want to re-use this message queueing for the other graphs instead of finding the node and queueing up the message directly on the node. Depends on our needs later.