Renamed most uses of Asset Script to Timeline.  This is because effectively the asset script is a timeline animation system utilising keyframes and command messages to control objects in game and will prevent potential ambiguity/confusion with script languages. Renamed TAssetScript.cpp and .h to TAssetTimeline.cpp and .h.
Made the TQuaternion GetEuler() routine const.
Added TQuaternion Nlerp and Slerp routines although they are not thoroughly tested yet.  If using these routines it is preferable to use the Nlerp rather Slerp.  TODO: Remove any redundant NLerp and Slerp routines as they vary a bit so need to narrow them down to one routine only.
Added TQuaternion DotProduct wrapper routine for use with Slerp.

Implemented the ability to create and shutdown nodes via the timeline system.  

Re-implemented the TAssetTimeline to now utilise a proper command reference with child data.  Previously the Data was used as a command ref.  This now means we can have commands with multiple data sent at the same time which was needed for creating nodes.  Transform messages now use a 'SetTransform' ref and contain translation, rotation and scale in one message rather than sending them as individual messages - should be slightly better than before.
NOTE: The XML format has changed so check the example files in the FindFritz project.

Updated the use of the "this" ref on the TTimelineInstance to use a generic keyarray as this will also be needed for tracking nodes that are created with differing TRef's to the ones specified in the XML data.

Due to the way the timeline command is a TBinaryTree message I've had to remove any local data as when it is exported/imported it uses the CopyDataTree which will overwrite any data previously in the tree on export and consequently import.  The only local data for a command was the Interpmethod which is now a piece of data attached to the command data instead.  The keyframe interping has changed to compensate.

Keyframe processing is now correct (going forward through the timeline) after extensive testing, and the last keyframe is processed as required.