Altered the graph and node ref to be optional when parsing the timeline XML.  This is so we can explicitly send messages to the subscribers of any timeline instances.
Added subscriber support to the TTimelineInstance class such that:

* when a timeline ends it will inform the subscribers with a 'OnComplete' message.
* when a timeline command is encoutnered with no node or graph ref it will be automatically passed onto the subscribers for processing

Implemented rudimentary timejump handling for the timelines so that we can jump to any time within the timeline by adding the following command:

<Command CommandRef="TimeJump">
     <Data DataRef="Time"><float>0</float></Data>
</Command>

This is demonstrate very simple looping of a timeline and future work in this area will involve possibly being able to jump to specific keyframes via a keyframe ref and tracking of nodes that are created and removed via the timeline within the time jump area.

Added initial work on a TSceneNode_Timeline so we can add a timeline node to the graph that would simply be there to play a timeline animation or sequence of commands when triggered to do so.  This would be used for playing static animations within a scene such as an object exploding or simple looped animations that can be easily setup via the scene XML (and eventually an editor) without having to add them in code.