Changed all uses of AddChildAndData to be ExportData instead as they do roughly the same thing.
I've left the original AddChildAndData call in but marked it deprecated.  

The only reason to use this call instead of the ExportData one is if you need a pointer to the TBinaryTree returning but I found only one instance of it's use in the engine which I've changed as it wasn't necessary and I suspect non of the games will need it either.