Added 8 bit colours types; TColour32 is an RGBA type and TColour24 is just RGB. Added 24 and 32 bit colour arrays to the vertex information in the meshes. (Will replace with a system where I can interlace the vertex data into one type next) Mesh's now use the 8bit versions when rendering (should save a good chunk of bandwidth when drawing). A flag exists if we want to explicitly use the floating point colours (better colour range) Moved all the vertex binding functions into generic code too as it was the same for PC and IPod.
Removed the IPod opengl check for error entirely as it's too slow to use and so just in case we can't ever call it now.

Added accessors for 24bit colour versions of textures.

Added ImportData and ExportData functions to TTransforms. These write valid (and optionally specific parts) of TTransforms to a TBinaryTree (eg. message). The read returns a bitmask of which parts were changed, and only changes stuff that has changed (i.e. even if we have a rotation in the message, if it hasn't changed it won't return the rotation bit). Changed a whole bunch of functions around to use the TTransform bitmasks instead of 3 bools.

I accidentally replaced all the inline's in the engine with FORCEINLINE, so some tabs might be a little off. Works still obviously.

Moved colour/component functions from TColour into a TLColour namespace. Renamed accessors to be more clear of the type and will allow templated use of the TColour* types. ALl types now implement GetRedf() and GetRef8() etc.

Changed the global core TRef's to use static versions. Changed some uses of these to be static versions instead of variable lookups.
Changed some other often-used ProcessMessage functions to be switch() statements too.

Added a macro to get a TShape type TRef in it's static form so we can use it for case statements and changed all the TShape's to use these. Good saving in the collision system!

Added a TLCounter namespace (currently in TLTime) like the timer counters to count stuff per-frame (e.g. collision tests)

Added missing != operators to some TypeX's.

Sped up the render node InvalidateBounds functions to do a lot less when they don't need to (should be a good saving) and speeds up the OnTransform bottleneck a bit.

Sped up the UpdateZone functions a bit, mostly in the form of an initial check now before trying to move the zone to see if it's possible to move at all. This means stuff where the physics/collision etc had been disabled no longer does a ton of redundant checks which would never succeed. Should be a good saving there. (Mostly for the physics post update, but applies to render and scene graph too)