Fixed issue with the g_TimerAverages array not being released properly. This is now emptied in the time manager shutdown.
Re-instated the deprecated on the PC.  When specifying a routine as deprecated put the define *before* the routine declaration:
   DEPRECATED void SomeRoutine();

This will now work with GCC and Microsoft's compiler.

Added cpuid.c and .h files for accessing processor information on a PC.
Moved the QueryHardwareInformation() and QueryLanguageInformation() calls so they are now done in the TCoreManager constructor.  This is because some of the information will at some stage be needed before the other managers are created and initialised.

Implemented the PC version of QueryHardwareInformation() and QueryLanguageInformation() and appropriate data is stored on the core manager.  In debug the processor and OS information will also be output to the console.

NOTE: On a PC most systems default user language is setup as US-EN which will be interpreted by the textmanager as "usa" whereas for the iphone there is only english for both US and UK (AFAIK) so when adding text specifically for a language in a text database XML file ensure you add both a usa and eng version at the very least for the text to appear.  Any specific languages not specified int he database will appear blank.

Added support for the <Text TextRef="Something" /> tag in the text database XML.  This will generate a text string the same as the text ref in all currently supported languages and can be used for quickly adding text without having to fill in the various localised versions.