Added Box2D (2.0.1) lib. Source, docs, examples and PC project.
Integrated Box2D into engine at the physics level only. Should be able to revert back to the old system by un-defining USE_BOX2D.

The QuadTree for the physics is no longer used (Not removed yet in case there is a benefit to using it in box2D for sweeps. Will decide when I have tested much larger worlds).
The WorldCollisionShape is now redundant. Will keep it in for now, but it will have much less practical use in future. There is still a benefit to having it (for rendering the shapes in world space) but it's not *needed* any more.

Fixed the rendering of transformed Box's in the physics-debug render node so they're transformed to TOblong's. This will move into the TShape code where it should be when I integrate intersection code from Box2D into the math lib. 

Added mesh-shape-creation from TOblong2Ds.

Added TTransform functions to set Translation/Rotation/Scale that return the transform bit only if the value has changed.

Corrected the TBox class's GetBoxCorner functions to return in Outline/Quad order. If you get any wierd trapezoid shapes generated instead of quads/boxes or similar then this is the cause.

Renamed the TransformVector function on TTransforms to just Transform().