Added box2D ipod project. Fixed compile error for GCC in box2D.
Added 64 bit colours, (RGBA u16 components) to see if the problem with the bad colours is a u8-specific issue on the ipod. It's probably something else, (Thought it might be colour conversion but then it should be broken on the PC too in that case) but added anyway.

Added a Debug scope timer (will compile out in release). Use in same way as normal one;
TLTime::Debug_ScopeTimer("Timer");

Added a TLCounter::Debug_Increment() which compiles out in release.

Corrected the maths and game ipod projects. The game one included itself in the sources. Maths I added new shape files (IIRC)

Added a custom contact filter for box2d to the physics graph. This should mean we do a few less body<->body collision tests. Can alter this further for more Tootle-specific collision-testing situations.
Added use of the GroupIndex in the body shapes so nodes set not to collide (eg. far-away cars) should no longer do collision tests.

Added an EnableCollision() function to the physics nodes which alters the flag but now does the box2d group-changing stuff

Added FORCE_COLOUR macro to TRenderTarget to hunt down this opengl render crash. Something wrong with the alternative colour buffer usage at the moment. Currently forcing to float colours which works fine. (But heavy on the GPU bus)

Scaled down the debug-text font in the screen's debug/stats render target when in landscape mode.