Added an "expected type" ref to TLAsset::LoadAsset so it will return NULL if not the desired type. (Only really works when the asset is loaded though, so good to use with the block load option)
Changed the Menu asset system to use "string" refs for text instead of "text" so it can switch over to use text refs (from text databases etc) for the "text" data. Makes menu -> render node translation easier. 

Added a block-load of audio assets when theyre attempted to be played in case they're not loaded. Slight bug though, even if I block load it, it doesn't play on that first play. But it does load the asset and subsequent plays DO make it play...

Added a check to RemoveNode in the TGraphs so we can catch cases where we've tried to delete a node without using a TPtr. The lack of referenced TPtr in the [temporarily] public RemoveNode function means it's easy to do. That then results in an object being deleted twice.

Added a rotation-minimum-change set function to TTransform to stop minor changes to rotations being passed around in physics changes.

Added a HasDifferenceMin() to the TypeX classes as a simple are-these-values-noticably different function.

Added a ortho-scalar XML option to SVG files; add this property to the <SVG> tag to make the width of the SVG document in 0-100 scale (screen scale). This makes it easier to make 2D/gui elements the right scale for the screen. Height of the document is irrelavant, but you'll have to make the ratio correct if you're explicitly setting the position for the screen.
TootleOrthoScale="TRUE"

Added XML data support for 24,32 and 64 bit colours.
<Col24>128,128,128</Col24> etc

Added a 2D d_circle asset to the virtual file sys.

Added a constructor to TWidgetButton so we can start using this class for buttons and not the base type.

Started moving bezier calculations to the maths lib from Tessellation (into TLine.h)

Transforming (with rotation) a 2D box shape now converts it into an oblong. Todo: remove the oblong shape and replace with the polygon type and use the box2d maths functions for intersection tests.

Added GetShapeFromBodyShape to convert box2d shape to a tootle shape (TShapePolygon2D)

Removed some non-box2D stuff for some speed up and removed some worldcollision shape generation in the physics node code. Added some body cleanup. Explicit changes to the transform now changes the box2d's body. 
Started expansion of the class to support multiple bodies and shapes. This was going in the TPhysicsNodeChain class, but may move to the base class as required. But keeping the base class as a single body/shape system might be best for us.

Added a TRenderNodeSwoosh type to render a line with variable widths at each point and smooth itself out (dynamically generated outsets etc) whilst still trying to be fast. Bezier support will be built in, but currently doesn't generate bezier steps.

Added support for arbritry strings back into TRenderNodeText. These are set with a "string" data ref. "Text" will be used for textref's instead.

Added a render flag for resetting the scene colour (ie. so a child doesn't inherit a parent's alpha or colour multiplier)

Added Init message merging (transform etc) to CreatePhysicsNode (same as CreateRenderNode) to SceneNode_Object. Also added ability to specify render node Type ref to CreateRenderNode to match CreatePhysicsNode.

Changed the SceneNode_Transform class so it doesn't sleep if it has no zone to cope with scenegraphs where we have no zone system.