Converted the mesh bounds types to TShape's instead of basic shape types. Start of change to make the bounds shapes just datums. Moved the bounds access around. There is no access to CalcBounds in TMesh's any more and instead GetBounds recalculates it as required. (usually only once in meshes) Same for rendernodes; Bounds shapes use TShapes now and will soon convert to be stored in a local & world datum system. The bounds shapes can now be accessed via a new GetLocalDatum() on a render node. If a non-bounds ref is passed in the render node looks up the datum in the mesh. World-translated support will be added as required.
Added the initial bits to setup text alignment and auto-scaling to fit into a text box. The text box can be taken from a datum in a mesh, or a datum in a render node (and in future will track the box as a render node moves around) or can be an explicit shape.
Only 2D boxes are supported at the moment (even if not implemented) but in future can be expanded to place text inside spheres etc.

Added a Shutdown() function to TStateMachines which should be called by anything that uses statemachines. This invokes an End() of the current mode (and goes to no mode). Many things use the mode end as a cleanup. TApplications and TGaames now shutdown their statemachines properly.

Added string "cleaning" for text databases and <Data> xml test. This converts \n to new lines and will do other cleanup and stuff later on as required. This may be an entry for runtime-string-replace code but need to think about that a bit more yet.

Changed the file sys factory's type creation to a switch statement. Also changed the recognisation of some file types to be a bit more readable. In future we should use .Timeline for a timeline instead of TTL. FNT will change to .Atlas or .Font or something. TTD is now .Text. TAM will change to .Markup etc

Added a CalcWorldTransform() function to TRenderNode's to calculate (if out of date) their world transform's. This is untested! Try to avoid using it, but I can imagine it will be handy when we want a world transform at Initialise of any object and the render node hasn't been rendered yet. The transform created should in theory match the post-render transform. It's just a little bit wasteful (but shouldn't be *too* expensive to use).