Corrected mesh bounds accessors to recalculate bounds when accessed (this was lost with the GCC compile fix)
Added an OnStringChanged function to TRenderNodeText's, will add a message to broadcast here when we need to know when a string changes on a render node (we may never need to know this :)

Added textbox support for text render nodes. Added alignement support too, so text can be placed Left,Center,Right with a HAlign ref data and Top, Middle, Bottom VAlign ref data.

Scale mode not-yet implemented.

This works on a whole block of text at the moment, if you have multiple lines they won't be individually centered/right aligned. I have something planned to fix this, but will probably require a bit more work. When it's implemented we will also be able to align text around 0,0,0 WITHOUT a box which can be handy.