Added more verbose XML errors. Added more verbose (and breaks) errors when files fail to load in TLoadTasks (much easier to track down problems now)
Added a TLFileSys::GetFileList() function to get a list of ALL files that exist in the file sys. (regardless of state/type/filesys)

Fixed the non-recognised file types to generate "asset" files, not "binary" files (an unhandled type). All the data in unrecognised files just go into the root of the TAsset's data.

Added STRef() macro to create static refs to replace TRef_Static. Not quite replaced yet, but the new macro is there.

Added a "sender" to the timeline "OnCompleted" message. The sender is the ref of the original timeline asset. This may need changing, but currently the timeline instances dont have individual ref's so this seemed appropriate. (but will probably need change as we re-use timelines I'd imagine)