Added an Initialise routine to the TTimelineInstance class.  You can now initialise the timeline starting time by passing a message with the required initial time exported to it.
Added a SetTime(float routine) routine to the TTimelineInstance class that can be used to manually set the time of a timeline.  This will trigger an immediate update of the timeline.

The TTimelineInstance class update now returns a SyncBool value.  If it returns SyncFalse the timeline has finished either naturally or via an error.  FYI SyncWait means the instance doesn;t have a valid asset - it might not be loaded yet, and SyncTrue signifies the update went through without error or reaching the end. 

Added a GetLastkeyframeTime to the TAssetTimeline class which simply returns the last keyframes time or zero if there are no keyframes.