Changed the generic data in the base TAsset type (Was used for storing lost data, stuff we don't read during import. Say, old data from a previous format or just plain data we don't import/export explicitly) Instead of an array of TBinaryTree's it's now a plain binary tree. Data should not be written to this (There is a break on export if there is data). Shouldn't affect anything as it's not really explicitly used yet, but I've incremented the .asset file version ref so all .asset files need to be rebuilt. (Should get an error as the .asset loads if it needs deleting.) I have also changed the sectioning header data structure (boring internal .asset file stuff) to save the data type ref. Which means old .asset files will be essentially corrupt anyway. This should be the last change I ever want to make to the .asset file format so from now on they will be backwards compatible. I might change it a little more so it's even more backwards compatible in the future, but first idea will bloat the file type a bit and slow down import...
Either way, all .asset files are now out of date!

The base TAsset type can now be imported & exported and be used for generic TBinaryTree data.

Changed OnAssetLoad to invalidate an asset (mark it as failed to load) and now deletes the associated asset if it's a Temp type (ie. asset file never existed) so that it CAN be created later as required.

Added a TLAsset::SaveAsset function which takes an asset, exports to a .asset file and then writes it to the file system.
This is currently explicitly used, but will later be a bit more automatic so when an asset changes it writes it's own changes back to the file system. (say if a mesh pre/calculates new colour data, bounds etc)

Added a catch for <data> xml import when some data has been declared, but is lacking any type information which previously resulted in a child data with no data inside. Usually just a result of mis-written scheme XML

Added quaternion (<quat>) data format support in XML. First use though seems wrong, but it may be more my usage or values than corruption. May need further testing.

When declaring a datum for a textbox from a render node you no longer need to specifiy a node. By default it will look in the parent node for a datum which was the most common usage.