Added work done so far on the small object allocator. This is essentially a cut-down version of the LOKI small object allocator.
Currently the SOA is disabled as it is still in testing.  There are two defines for testing the SOA system:

* ENABLE_SOA is used to enable the SOA system itself
* TEST_COUNTER_USING_SOA is used to test the SOA using the TPtr's counter object.  This object is allocated when a new TPtr is instanced (which is often) so it is a good starting point for ensuring the system works.

I'm committing the code so that I have a backup of it and in case you want to have a play with it at any stage Graham.

In addition to the SOA I have added some initial work on getting an alert window appearing on the iPhone when a debug break is encountered - mainly because I was getting annoyed with the break not being handled at all properly on the iphone for little things like assets missing that result in searching the text output in debug and takes ages.  

This works OK in so much as the alert will appear and show the error message as text BUT we really need it to stop in the TLDebug_Break until we know the result of the users button selection to determine whether (to break or continue) which unfortunately doesn't appear possible from my investigations as the alert isn;t a modal object so we can't sit in a loop and wait for the alert to return a result. Therefore this is also disabled whilst I look into it further.