Added string-replace functionality to strings. Add {Ref} to a string in text database XML to create a "replace" entry. Then when you request a string with g_TextManager.GetString() provide a key-array of ref's with strings to insert at the correct place.
I've also cut down the strings stored in the text assets to buffer strings (limited to 255 chars). I don't think we're going to need more than this. If we do need more then we can just replace the type in TComplexString and not have to do anything else. 

The text asset type's data has changed to. The new data format is now backwards compatible... but not with the previous structure. Delete the old text asset files if you get an error. The XML hasn't changed.

Added a check in TBinaryTree::Debug_PrintTree to stop a crash.

Added an asset factory shutdown, but still crashes on PC app exit...