Added new TWidgetText.cpp and .h files to the TootleGame iPod project. Implemented the text widget handling which will create the virtual keyboard device when clicked on and intercept the 'return' button from the keyboard to remove the virtual device.
When the device is created or removed the TWidgetText will receive the device added and removed notifications and update the global users action maps with the required actions used for each key.  The input system has a ref -> character map that is used to determine the character to print when a key is pressed and this map is used as a look up when a supported key is pressed on the virtual keyboard.  This may need further work for different languages and/or character formats i.e. unicode or UTF8. 

Because the widget handles the virtual device creation and removal tied to a specific text render node we *should* (although not tested) be able to have a set of text boxes with text widgets and each individual one will be able to be edited separately. See the KeyboardTest files in the test project for usage examples.

Moved the CreateVirtualDevice and RemoveVirtualDevice calls onto the input manager and made them async requests (only one frame) rather than immediate routines.  This prevents issues with trying to remove a virtual device from a callback from the device itself and makes it consistent with other parts of the engine.

Changed the CreateVirtualDevice to go through the standard input device creation handling.  You can now also specify they type of virtual device to create although at the moment only a virtual keyboard is supported.  This means all the device added message will be triggered when creating a virtual device just like with a physical device and you can map actions to the device as well.

Updated the RemoveVirtualDevice to send out a message of it's removal prior to removing the virtual device.

Added a keyboard ref map for the iPod keyboard device.  This works in a similar way to the PC version where we have a key ref which then corresponds to a gameside label we can use for the key 'sensor' objects. On the PC the key is the DirectX DIK enum values and on the iPod they are a TRef created from the character for the key - this is because on the iPod we don't have an equivalent DIK type enum corresponding to the keyboard 'button' so rather than do extra conversions we use the character passed to us as the key directly.  This should work transparently but *may* cause issues with different language keyboards which will need testing, but should simply involve adding all possible characters that are keys on the iPod keyboard. 

The iPod virtual keyboard device now processes the return key and backspace key as special cases.  A 'k_backspace' action will trigger for the backspace button and 'k_return' for the return/done/send button.

Disabled the touch events whilst the virtual keyboard is active.

Added a SetString message to the TRenderNodeText so I can simply send it a setstring message rather than going through the initialise multiple times.

Added a DISABLE_ACCELEROMETER define to be able to enable/disable the accelerometer - mainly for testing to remove the large number of traces in debug.