Added the TWidgetText .cpp and .h files to the PC iPod TootleGame library.
Moved the OnDeviceAdded and OnDeviceRemoved handling from the TWidgetText tot he TWidgetManager.  This is bacuase the actions that will be mapped to the keyboard keys don;t change and it ensures the device change interception for the keyboard will work on both PC and iPod without adding any custom device lookup if the creation of the virtual device fails. It also simplifies the code slightly reducing the number of objects that would otherwise intercept this message.

Removed the Intiialise and Shutdown calls from the TWidgetText as they aren't really needed.