Added a 'Publish Feed' button to the facebook test. When clicked on this will bring up the facebook feed dialogue for posting a feed with the feed information specified for the facebook test application - setup on the facebook pages.
Click the dialogues 'publish' button to see the info posted onto your profile.  The feed is not publicly visible due to the facebook test applicaiton being a development application.

Added a 'Get permission' button to setup the permissions for automatically posting feeds without having to ask the user. This is just a demo of how to do this - we don't have any automated type feed example yet.