Added an ActionRef to the Click object class. This is used to know what action triggered the 'click' so we can determine if it was a click or move event. Also updated virtual OnCursorMove and QueueClick routines to use the actionref.
Removed the TInputInterface action creation for all 'mouse' type buttons per input interface object in the initialise routine. Instead a set of generic actions for all 'mouse' type buttons is created by the widget manager reducing down the number of actions intercepting messages to a finite amount. The widget manager keeps track of the action ref's created and you can test to see if they are click actions or move actions if required.

Added an 'action being processed' TRef property to the TInputInterface class. This is used to track a click action specifically and fixes an issue with multiple clicks on the same object (rocket circle) toggling states as well as clicks on different TInputInterface objects (Hockey) toggling different object states.