Moved the base-base graph node initialise to the TGraphNodeBase class. The initialise here now copies any unread data from the init message into the node's NodeData (which is now publiccly accessible). This means any arbritry data put into a graph node will be saved in the "reflection" data. Similar to the asset system.
Added GetNearestNodes to the physics graph. Given a shape this returns a list of physics nodes [shapes] that are inside (intersect I think) that shape.

Added "object" factory entry to the scenegraph so we can create simple scene node objects.

Added Initialise functionality to SceneNode_Object::Initialise to auto-create render nodes and physics node.
If a mesh ref or specific render node type is specified then a render node is created.
If a colshape (shape data) or coldatum (datum in the specified meshref) then a physics node is created. 
All the data in the SceneNodeObject init message is passed onto the auto-created rendernode and/or physics nodes, so you can add physics/render flags etc in the scene init