Changed TBinaryTree to return a TPtr reference to the data added/found/etc with ImportData/ExportData etc. Will be a tiny tiny minor hit to performance to return a pointer instead of a bool. Existing code needs no changes unless it was checking for ==/!= TRUE/FALSE instead of just an if ()
Removed use of my box2d local change of InitializeLocal for the shapes which was a bit of a speed up. Trying to update box2d to SVN version to make use of some new features so local changes have to go.

Added a "Force" message handling for physics nodes to send it a force vector (float3) to apply a force remotely.
Added a "ResetForces" message which stops all velocity/forces.

Removed default anti-aliasing so PC will look more like the ipod.

SceneNode_Objects now handle an "enable" message which has children to enable/disable the render and physics nodes remotely.