Added Freeze/Unfreeze functionality to box2D. Added a patch file for this if we need to re-implement it (and I've submitted this patch to box2D)
Removed the TShapeOblong type and replaced it's use with TShapePolygon2D. The basic TLMaths::TOblong2D will be replaced with a polygon equivalent soon.

Removed debug counters from physicsgraph. If we're to reimplement any we should use the debug counters.
Removed quadtree integration from the physics graph and physics nodes. This allowed me to removed the "world collision shape" from the physics node. There is a function to fetch all of the shapes attached to a body in world space instead. This will be slower, but currently the only need for this is debug rendering.
Removed all the old collision testing code now that the zones are gone. More old [2d] collision maths to go soon as I integrate the box2D math as required.

Fixed a node's body transform being not updated when the node's transform was changed. (bug was from Init message but fix is global)

Physics node's body is now frozen/unfrozen when the node is enabled/disabled so it won't be doing any position iterations (ie. gravity, velocity etc won't be moving it by box2D when disabled)