Added the code to the generic GenerateShape() function for the mesh which does a switch so the right kind of geometry is created.
Added extra geometry generation code for mesh. Currently polygons with <> 4 points are turned into outlines until I've integrated/replaced the tesselator.

Fixed TLPhysics::GetShapeFromBodyShape so that it uses the NODE's transform to transform a body's shape into world space. If we use box2D's transform when the node/body is disabled/frozen then these shapes are transformed by the body's transform which is out of date as we cannot set a transform on a body when it's frozen.
This is mostly a fix for the debug-rendering of a physics node (eg. in the test game you wouldnt see the object move when you dragged it even though the transform was correct). But for any future use of this function (which isn't a debug function) then the transforms of the shapes will be accurate according to the node.

Added a DoTransform message handling for the render node (should probably copy this to scene & physics nodes).
This transforms the current Node Transform by the Transform in the message. ie. Moves a node. 
Cleaned up the SetTransform message too.