Removed the oblong CreateDatum code on the mesh as it's the same as the polygon shape generation. (Should probably move this into generic maths CreateShapeFromPoints code)
Added a simple TMesh::GetDatumCenter() function to the get the center of a datum.

Changed TShapePolygon to clean up the contour when it changes. This now ensures there are no overlapping points, and the shape is always clockwise. This code will move into a simplified maths polygon shape and out of the TShape code when that class is implemented.
This is mostly to fix up errors when generating datums for use in box2d from an SVG