Fixed the issue with the quaternion slerp and nlerp not working correctly - this was down to some of the Type4<> operators not returning an explicit Type4<> object so some temporary variables would trigger the Type4<> constructor with a single variable rather than a Type4<> variable, therefore the Quaternions float4 would be incorrect when using these operators.  Check changes to TLTypes.inc.h to see modifications.
Also corrected issues with the NLerp and SLerp routines themselves now that they are working, such as selecting the longest route for interp rather than the shortest.  Check the animation test in the TestGame to see all versions working as required.

Added support for <Euler> and <AxisAngle> properties in the XML files.  In addition, I've added two convenince maths classes TEuler and TAxisAngle for easy setup and association with these properties so you can use the following to check the property type in the XML:

* TLBinary::GetDataTypeRef<TLMaths::TEuler>()) - check for the <Euler> property
* TLBinary::GetDataTypeRef<TLMaths::TAxisAngle>()) - check for the <AxisAngle> property

Fixed an issue with the TLMaths::TQuaternion::SetEuler() using the X axis for the Z axis and vice versa.

Fixed an issue with the TFileTimeline not exporting the InterpMethod or Mode correctly.

Added support to the TFileTimeline for the new <Euler> and <AxisAngle> properties. When used these will trigger the correct conversion to a quaternion for the rotations.  <float3> is still supported and assumed to be the same as Euler angles.  <Quat> is also now supported for the rotations which will import the raw quaternion data.  At the moment all angles, other than for the <Quat>, are assumed to be in degrees (mainly because it is easier to work in degrees when editing manually) but support for radians will be added soon.  This will either be via an 'AngleFormat' property for the data or by using the <float3> and <float4> as radian versions and leaving the <AxisAngle> and <Euler> as degree based versions.

Updated the TTimeline to detect the InterpMethod correctly for rotations and use either the NLerp or SLerp routines as required.